+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/RobotHighlighter.java
@@ -13,35 +13,114 @@ import java.util.Map;
 
 public class RobotHighlighter extends SyntaxHighlighterBase {
 
-    private static final Map<IElementType, TextAttributesKey> keys1 = new HashMap<>();
-
-    public static final TextAttributesKey HEADING = TextAttributesKey.createTextAttributesKey("Section Heading", DefaultLanguageHighlighterColors.STRING);
-    public static final TextAttributesKey SETTING = TextAttributesKey.createTextAttributesKey("Global Setting Option",
-                                                                                              DefaultLanguageHighlighterColors.DOC_COMMENT_TAG_VALUE);
-    public static final TextAttributesKey LOCAL_SETTING = TextAttributesKey.createTextAttributesKey("Local Setting Option",
-                                                                                                    DefaultLanguageHighlighterColors.DOC_COMMENT_TAG);
-    public static final TextAttributesKey IMPORT = TextAttributesKey.createTextAttributesKey("Import Option",
+    public static final TextAttributesKey SECTION_TITLE = TextAttributesKey.createTextAttributesKey("ROBOT_SECTION_TITLE",
+                                                                                                    DefaultLanguageHighlighterColors.STRING);
+    public static final TextAttributesKey GLOBAL_SETTING_OPTION = TextAttributesKey.createTextAttributesKey("ROBOT_GLOBAL_SETTING_OPTION",
+                                                                                                            DefaultLanguageHighlighterColors.DOC_COMMENT_TAG_VALUE);
+    public static final TextAttributesKey LOCAL_SETTING_OPTION = TextAttributesKey.createTextAttributesKey("ROBOT_LOCAL_SETTING_OPTION",
+                                                                                                           DefaultLanguageHighlighterColors.DOC_COMMENT_TAG);
+    public static final TextAttributesKey IMPORT = TextAttributesKey.createTextAttributesKey("ROBOT_IMPORT_OPTION",
                                                                                              DefaultLanguageHighlighterColors.DOC_COMMENT_TAG_VALUE);
-    public static final TextAttributesKey IMPORT_ARGUMENT = TextAttributesKey.createTextAttributesKey("Import Argument",
+    public static final TextAttributesKey IMPORT_ARGUMENT = TextAttributesKey.createTextAttributesKey("ROBOT_IMPORT_ARGUMENT",
                                                                                                       DefaultLanguageHighlighterColors.HIGHLIGHTED_REFERENCE);
-    public static final TextAttributesKey KEYWORD_DEFINITION = TextAttributesKey.createTextAttributesKey("User Keyword Definition",
+    public static final TextAttributesKey KEYWORD_DEFINITION = TextAttributesKey.createTextAttributesKey("ROBOT_USER_KEYWORD_DEFINITION",
                                                                                                          DefaultLanguageHighlighterColors.KEYWORD);
-    public static final TextAttributesKey KEYWORD = TextAttributesKey.createTextAttributesKey("Keyword Usage",
+    public static final TextAttributesKey KEYWORD = TextAttributesKey.createTextAttributesKey("ROBOT_KEYWORD_CALL",
                                                                                               DefaultLanguageHighlighterColors.FUNCTION_DECLARATION);
-    public static final TextAttributesKey PARAMETER = TextAttributesKey.createTextAttributesKey("Parameter", DefaultLanguageHighlighterColors.PARAMETER);
-    public static final TextAttributesKey ARGUMENT = TextAttributesKey.createTextAttributesKey("Positional Argument",
+    public static final TextAttributesKey PARAMETER = TextAttributesKey.createTextAttributesKey("ROBOT_PARAMETER", DefaultLanguageHighlighterColors.PARAMETER);
+    public static final TextAttributesKey ARGUMENT = TextAttributesKey.createTextAttributesKey("ROBOT_POSITIONAL_ARGUMENT",
                                                                                                DefaultLanguageHighlighterColors.STATIC_FIELD);
-    public static final TextAttributesKey VARIABLE = TextAttributesKey.createTextAttributesKey("Variable Usage",
+    public static final TextAttributesKey VARIABLE = TextAttributesKey.createTextAttributesKey("ROBOT_VARIABLE",
                                                                                                DefaultLanguageHighlighterColors.DOC_COMMENT_MARKUP);
-    public static final TextAttributesKey VARIABLE_DEFINITION = TextAttributesKey.createTextAttributesKey("Variable Definition",
+    public static final TextAttributesKey VARIABLE_DEFINITION = TextAttributesKey.createTextAttributesKey("ROBOT_VARIABLE_DEFINITION",
                                                                                                           DefaultLanguageHighlighterColors.DOC_COMMENT_MARKUP);
-    public static final TextAttributesKey EXTENDED_VARIABLE_ACCESS_BRACKETS = TextAttributesKey.createTextAttributesKey("Extended Variable Brackets",
-                                                                                                                        DefaultLanguageHighlighterColors.BRACKETS);
-    public static final TextAttributesKey COMMENT = TextAttributesKey.createTextAttributesKey("Comment", DefaultLanguageHighlighterColors.LINE_COMMENT);
-    public static final TextAttributesKey GHERKIN = TextAttributesKey.createTextAttributesKey("Gherkin Keywords", DefaultLanguageHighlighterColors.METADATA);
-    public static final TextAttributesKey SYNTAX_MARKER = TextAttributesKey.createTextAttributesKey("Robot Keyword", DefaultLanguageHighlighterColors.METADATA);
+    public static final TextAttributesKey EXTENDED_VARIABLE_ACCESS_BRACKETS = TextAttributesKey.createTextAttributesKey(
+            "ROBOT_EXTENDED_VARIABLE_ACCESS_BRACKETS",
+            DefaultLanguageHighlighterColors.BRACKETS);
+    public static final TextAttributesKey COMMENT = TextAttributesKey.createTextAttributesKey("ROBOT_COMMENT", DefaultLanguageHighlighterColors.LINE_COMMENT);
+    public static final TextAttributesKey GHERKIN = TextAttributesKey.createTextAttributesKey("ROBOT_GHERKIN_KEYWORD",
+                                                                                              DefaultLanguageHighlighterColors.METADATA);
+    public static final TextAttributesKey STRUCTURAL_KEYWORDS = TextAttributesKey.createTextAttributesKey("ROBOT_STRUCTURAL_KEYWORD",
+                                                                                                          DefaultLanguageHighlighterColors.METADATA);
     public static final TextAttributesKey ERROR = TextAttributesKey.createTextAttributesKey(TokenType.BAD_CHARACTER.toString(),
                                                                                             DefaultLanguageHighlighterColors.INVALID_STRING_ESCAPE);
+
+    private static final Map<IElementType, TextAttributesKey> keys = new HashMap<>();
+    static {
+        keys.put(RobotTypes.SETTINGS_HEADER, SECTION_TITLE);
+        keys.put(RobotTypes.VARIABLES_HEADER, SECTION_TITLE);
+        keys.put(RobotTypes.USER_KEYWORDS_HEADER, SECTION_TITLE);
+        keys.put(RobotTypes.COMMENTS_HEADER, SECTION_TITLE);
+        keys.put(RobotTypes.TEST_CASES_HEADER, SECTION_TITLE);
+        keys.put(RobotTypes.TASKS_HEADER, SECTION_TITLE);
+
+        keys.put(RobotTypes.USER_KEYWORD_NAME, KEYWORD_DEFINITION);
+        keys.put(RobotTypes.TEST_CASE_NAME, KEYWORD_DEFINITION);
+        keys.put(RobotTypes.TASK_NAME, KEYWORD_DEFINITION);
+
+        keys.put(RobotTypes.COMMENT, COMMENT);
+        keys.put(RobotTypes.PARAMETER_NAME, PARAMETER);
+        keys.put(RobotTypes.TEMPLATE_PARAMETER_NAME, PARAMETER);
+        keys.put(RobotTypes.LITERAL_CONSTANT, ARGUMENT);
+        keys.put(RobotTypes.TEMPLATE_ARGUMENT_VALUE, ARGUMENT);
+
+        keys.put(RobotTypes.GIVEN, GHERKIN);
+        keys.put(RobotTypes.WHEN, GHERKIN);
+        keys.put(RobotTypes.THEN, GHERKIN);
+        keys.put(RobotTypes.AND, GHERKIN);
+        keys.put(RobotTypes.BUT, GHERKIN);
+
+        keys.put(RobotTypes.FOR, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.FOR_IN, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.WHILE, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.TRY, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.EXCEPT, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.FINALLY, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.RETURN, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.GROUP, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.IF, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.ELSE_IF, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.ELSE, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.END, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.BREAK, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.CONTINUE, STRUCTURAL_KEYWORDS);
+        keys.put(RobotTypes.WITH_NAME, STRUCTURAL_KEYWORDS);
+
+        keys.put(RobotTypes.VARIABLE_DEFINITION, VARIABLE_DEFINITION);
+        keys.put(RobotTypes.SCALAR_VARIABLE_START, VARIABLE);
+        keys.put(RobotTypes.LIST_VARIABLE_START, VARIABLE);
+        keys.put(RobotTypes.DICT_VARIABLE_START, VARIABLE);
+        keys.put(RobotTypes.ENV_VARIABLE_START, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_BODY, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_END, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_KEY_ACCESS, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_SLICE_ACCESS, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_INDEX_ACCESS, VARIABLE);
+        keys.put(RobotTypes.VARIABLE_ACCESS_START, EXTENDED_VARIABLE_ACCESS_BRACKETS);
+        keys.put(RobotTypes.VARIABLE_ACCESS_END, EXTENDED_VARIABLE_ACCESS_BRACKETS);
+        keys.put(RobotTypes.USER_KEYWORD_STATEMENT, KEYWORD_DEFINITION);
+        keys.put(RobotTypes.KEYWORD_LIBRARY_NAME, KEYWORD);
+        keys.put(RobotTypes.KEYWORD_LIBRARY_SEPARATOR, KEYWORD);
+        keys.put(RobotTypes.KEYWORD_NAME, KEYWORD);
+
+        keys.put(RobotTypes.SUITE_NAME_KEYWORD, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.DOCUMENTATION_KEYWORD, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.METADATA_KEYWORD, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.SETUP_TEARDOWN_STATEMENT_KEYWORDS, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.TAGS_KEYWORDS, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.TEMPLATE_KEYWORDS, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.TIMEOUT_KEYWORDS, GLOBAL_SETTING_OPTION);
+        keys.put(RobotTypes.UNKNOWN_SETTING_KEYWORD, GLOBAL_SETTING_OPTION);
+
+        keys.put(RobotTypes.LOCAL_SETTING_NAME, LOCAL_SETTING_OPTION);
+        keys.put(RobotTypes.ARGUMENTS_SETTING_NAME, LOCAL_SETTING_OPTION);
+
+        keys.put(RobotTypes.LIBRARY_IMPORT_KEYWORD, IMPORT);
+        keys.put(RobotTypes.RESOURCE_IMPORT_KEYWORD, IMPORT);
+        keys.put(RobotTypes.VARIABLES_IMPORT_KEYWORD, IMPORT);
+
+        keys.put(TokenType.BAD_CHARACTER, ERROR);
+    }
 
     @NotNull
     @Override
@@ -49,85 +128,9 @@ public class RobotHighlighter extends SyntaxHighlighterBase {
         return new RobotLexerAdapter();
     }
 
-    static {
-        keys1.put(RobotTypes.SETTINGS_HEADER, HEADING);
-        keys1.put(RobotTypes.VARIABLES_HEADER, HEADING);
-        keys1.put(RobotTypes.USER_KEYWORDS_HEADER, HEADING);
-        keys1.put(RobotTypes.COMMENTS_HEADER, HEADING);
-        keys1.put(RobotTypes.TEST_CASES_HEADER, HEADING);
-        keys1.put(RobotTypes.TASKS_HEADER, HEADING);
-
-        keys1.put(RobotTypes.USER_KEYWORD_NAME, KEYWORD_DEFINITION);
-        keys1.put(RobotTypes.TEST_CASE_NAME, KEYWORD_DEFINITION);
-        keys1.put(RobotTypes.TASK_NAME, KEYWORD_DEFINITION);
-
-        keys1.put(RobotTypes.COMMENT, COMMENT);
-        keys1.put(RobotTypes.PARAMETER_NAME, PARAMETER);
-        keys1.put(RobotTypes.TEMPLATE_PARAMETER_NAME, PARAMETER);
-        keys1.put(RobotTypes.LITERAL_CONSTANT, ARGUMENT);
-        keys1.put(RobotTypes.TEMPLATE_ARGUMENT_VALUE, ARGUMENT);
-
-        keys1.put(RobotTypes.GIVEN, GHERKIN);
-        keys1.put(RobotTypes.WHEN, GHERKIN);
-        keys1.put(RobotTypes.THEN, GHERKIN);
-        keys1.put(RobotTypes.AND, GHERKIN);
-        keys1.put(RobotTypes.BUT, GHERKIN);
-
-        keys1.put(RobotTypes.FOR, SYNTAX_MARKER);
-        keys1.put(RobotTypes.FOR_IN, SYNTAX_MARKER);
-        keys1.put(RobotTypes.WHILE, SYNTAX_MARKER);
-        keys1.put(RobotTypes.TRY, SYNTAX_MARKER);
-        keys1.put(RobotTypes.EXCEPT, SYNTAX_MARKER);
-        keys1.put(RobotTypes.FINALLY, SYNTAX_MARKER);
-        keys1.put(RobotTypes.RETURN, SYNTAX_MARKER);
-        keys1.put(RobotTypes.GROUP, SYNTAX_MARKER);
-        keys1.put(RobotTypes.IF, SYNTAX_MARKER);
-        keys1.put(RobotTypes.ELSE_IF, SYNTAX_MARKER);
-        keys1.put(RobotTypes.ELSE, SYNTAX_MARKER);
-        keys1.put(RobotTypes.END, SYNTAX_MARKER);
-        keys1.put(RobotTypes.BREAK, SYNTAX_MARKER);
-        keys1.put(RobotTypes.CONTINUE, SYNTAX_MARKER);
-        keys1.put(RobotTypes.WITH_NAME, SYNTAX_MARKER);
-
-        keys1.put(RobotTypes.VARIABLE_DEFINITION, VARIABLE_DEFINITION);
-        keys1.put(RobotTypes.SCALAR_VARIABLE_START, VARIABLE);
-        keys1.put(RobotTypes.LIST_VARIABLE_START, VARIABLE);
-        keys1.put(RobotTypes.DICT_VARIABLE_START, VARIABLE);
-        keys1.put(RobotTypes.ENV_VARIABLE_START, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_BODY, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_END, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_KEY_ACCESS, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_SLICE_ACCESS, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_INDEX_ACCESS, VARIABLE);
-        keys1.put(RobotTypes.VARIABLE_ACCESS_START, EXTENDED_VARIABLE_ACCESS_BRACKETS);
-        keys1.put(RobotTypes.VARIABLE_ACCESS_END, EXTENDED_VARIABLE_ACCESS_BRACKETS);
-        keys1.put(RobotTypes.USER_KEYWORD_STATEMENT, KEYWORD_DEFINITION);
-        keys1.put(RobotTypes.KEYWORD_LIBRARY_NAME, KEYWORD);
-        keys1.put(RobotTypes.KEYWORD_LIBRARY_SEPARATOR, KEYWORD);
-        keys1.put(RobotTypes.KEYWORD_NAME, KEYWORD);
-
-        keys1.put(RobotTypes.SUITE_NAME_KEYWORD, SETTING);
-        keys1.put(RobotTypes.DOCUMENTATION_KEYWORD, SETTING);
-        keys1.put(RobotTypes.METADATA_KEYWORD, SETTING);
-        keys1.put(RobotTypes.SETUP_TEARDOWN_STATEMENT_KEYWORDS, SETTING);
-        keys1.put(RobotTypes.TAGS_KEYWORDS, SETTING);
-        keys1.put(RobotTypes.TEMPLATE_KEYWORDS, SETTING);
-        keys1.put(RobotTypes.TIMEOUT_KEYWORDS, SETTING);
-        keys1.put(RobotTypes.UNKNOWN_SETTING_KEYWORD, SETTING);
-
-        keys1.put(RobotTypes.LOCAL_SETTING_NAME, LOCAL_SETTING);
-        keys1.put(RobotTypes.ARGUMENTS_SETTING_NAME, LOCAL_SETTING);
-
-        keys1.put(RobotTypes.LIBRARY_IMPORT_KEYWORD, IMPORT);
-        keys1.put(RobotTypes.RESOURCE_IMPORT_KEYWORD, IMPORT);
-        keys1.put(RobotTypes.VARIABLES_IMPORT_KEYWORD, IMPORT);
-
-        keys1.put(TokenType.BAD_CHARACTER, ERROR);
-    }
-
     @NotNull
     @Override
     public TextAttributesKey @NotNull [] getTokenHighlights(IElementType tokenType) {
-        return pack(keys1.get(tokenType));
+        return pack(keys.get(tokenType));
     }
 }